+++ src/services/tts/index.ts
@@ -1,6 +1,7 @@
 export * from './constants';
 export * from './types';
-export {SystemEngine} from './SystemEngine';
-export {SupertonicEngine} from './SupertonicEngine';
+export {SystemEngine} from './engines/system';
+export {SupertonicEngine} from './engines/supertonic';
+export {SUPERTONIC_VOICES} from './engines/supertonic/voices';
 export {getEngine, getAllEngines} from './engineRegistry';
 export {configureAudioSession} from './audioSession';
